# Optimize isSharingDisabledForUser