--- a/lib/private/Share20/Manager.php
+++ b/lib/private/Share20/Manager.php
@@ -1212,6 +1212,10 @@ class Manager implements IManager {
 	 * @return bool
 	 */
 	public function sharingDisabledForUser($userId) {
+		if ($userId === null) {
+			return false;
+		}
+
 		if (isset($this->sharingDisabledForUsersCache[$userId])) {
 			return $this->sharingDisabledForUsersCache[$userId];
 		}