I was doing some benchmarking and profiling for the common propfinds of the client. And on a directory with 20K entries I noticed a significant impact of the isSharingDisabledForUser function.

[Blackfire graph](https://blackfire.io/profiles/compare/939ac718-4d11-4d2a-8c7a-87ce89b6b13c/graph)

Our own perf tool (https://github.com/owncloud/administration/tree/master/performance-tests-c%2B%2B) also show improvements for the propfinds:

  |master | branch
-----------------|----------|-----------
RequestEtagJob | 251.1 ms | 234.1 ms
ConnectionValidator | 267.1 ms | 256.3 ms
DiscoverySingleDirectoryJob | 612.8 ms | 529.0 ms
CheckQuotaJob | 253.0 ms | 246.3 ms

Eventaully we should properly inject the sharemanager. But this was already such a nice preformance improvement.

CC: @icewind1991 @PVince81 @MorrisJobke @nickvergessen