@rullzer 2016-05-03
@PVince81 well 20K entries is a lot... but the saving is also significant...