@rullzer 2016-05-03
Aaah mmm it seems I broke a few tests... let me fix that tomorrow... I think it is because it are semi intergration tests so we cache the result but don't reset the cache..