@rullzer 2016-06-14
OK unit tests pass now at least locally...
review time!